Renaming 'NANO_Compat.h/cpp' to 'NANO_compat.h/cpp' for the naming scheme to be identical to 'UNO_compat.h/cpp'
