+++ cores/arduino/Arduino.h
@@ -144,7 +144,7 @@ bool isDoubleBondedActive(uint8_t pin);
 #include <UNO_compat.h>
 #endif
 #ifdef AVR_NANO_4809_328MODE
-#include <NANO_Compat.h>
+#include <NANO_compat.h>
 #endif
 #if defined(HAVE_HWSERIAL0) && defined(HAVE_CDCSERIAL)
 #error "Targets with both UART0 and CDC serial not supported"
