+++ cores/arduino/NANO_compat.h
